# [LI-HOTFIX] Replace Java stream api usages with non-functional old style in ZkMetadata request/response path