EXIT_CRITERIA = After KIP-500 and we're not using ZK for metadata

Changes
=======

This is a combination porting of the 2 commits from 2.4-li branch to 3.0:
- def7e89 Replace java.util.stream.* api usages with non-functional old style code in Metadata request/response (#163)
- 5b170af [LI-HOTFIX] Fix a spotbugs warning in JMH benchmark code (#194)

- Replaced java.util.stream.* api usages with non-functional old style code in
  - `MetadataRequest`
  - `MetadataResponse`
  - `MetadataCache`
- Added a JMH benchmark code for `MetadataResponse.prepareResponse` codepath.

Tests
=====

In the original 2.4 PR (linkedin#163), the benchmark shows ~3x improvement in GC allocation rates for Metadata request/response.
However, after porting to the 3.0 branch, the improvment is not that significant or even worse.

**Baseline**
```
MetadataRequestResponseBenchmark.responseConstructor:·gc.alloc.rate.norm               thrpt   10  824.005 ±  0.001    B/op
MetadataRequestResponseBenchmark.responseConstructor:·gc.time                          thrpt   10  885.000               ms
```

**Test**
```
MetadataRequestResponseBenchmark.responseConstructor:·gc.alloc.rate.norm               thrpt   10   904.005 ±  0.001    B/op
MetadataRequestResponseBenchmark.responseConstructor:·gc.time                          thrpt   10   874.000               ms
```

Minor (< 1%) improvements in `ZkMetadataCache` GC allocation rates.
Same as the original test results in 2.4

**Baseline**
```
ZkMetadataCacheBenchmark.benchmarkGetTopicMetadata:·gc.alloc.rate.norm               thrpt   10   960.000 ±  0.001    B/op
ZkMetadataCacheBenchmark.benchmarkGetTopicMetadata:·gc.time                          thrpt   10  5187.000               ms
```

**Test**
```
ZkMetadataCacheBenchmark.benchmarkGetTopicMetadata:·gc.alloc.rate.norm               thrpt   10   840.000 ±  0.001    B/op
ZkMetadataCacheBenchmark.benchmarkGetTopicMetadata:·gc.time                          thrpt   10  5269.000               ms
```