@lmr3796 2022-03-10
Talked with @gitlw :

Since we're not seeing significant improvements after bringing to 3.0, we can discard the patch.